implemented tempus edge message handling in opcClient
