+++ applications/opc-client/src/main/scala/com/hashmapinc/tempus/edge/opcClient/iofog/IofogConnection.scala
@@ -18,7 +18,7 @@ object IofogConnection {
    */
   def requestConfigs: Unit = {
     log.info("Requesting config from iofog...")
-    client.fetchContainerConfig(IofogController)
+    client.fetchContainerConfig(IofogListener)
   }
   
   /**
@@ -27,12 +27,12 @@ object IofogConnection {
   def connect: Unit = {
     log.info("Creating iofog connection")
     try {
-      client.openMessageWebSocket(IofogController)
+      client.openMessageWebSocket(IofogListener)
     } catch {
       case e: Exception => log.error("IoFog websocket error: " + e.toString)
     }
     try {
-      client.openControlWebSocket(IofogController)
+      client.openControlWebSocket(IofogListener)
     } catch {
       case e: Exception => log.error("IoFog websocket error: " + e.toString)
     }
